rag_app/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
jinja2==3.1.6
numpy==2.3.2
faiss-cpu==1.12.0
sentence-transformers==5.1.0
openai==1.107.2
pypdf==5.9.0
